+++ booking-admin/src/app/(protected)/bookings/_components/interactive-guest-list/index.tsx
@@ -14,24 +14,7 @@ import { useEffect, useState } from "react";
 import { strapiService } from "@/utils/dataEndpoint";
 import UpdateCustomerBooking from "../update-booking-modal";
 import React from "react";
-
-const generatePastWeekDateRanges = () => {
-  const now = new Date();
-  const ranges = [];
-  for (let i = 0; i <= 7; i++) {
-    const pastDate = new Date(now.getTime() - i * 24 * 60 * 60 * 1000);
-    ranges.push({
-      label:
-        i === 0
-          ? `Today (${pastDate.toLocaleDateString()})`
-          : i === 1
-          ? `Yesterday (${pastDate.toLocaleDateString()})`
-          : `${i} days ago (${pastDate.toLocaleDateString()})`,
-      value: pastDate.toISOString().split("T")[0],
-    });
-  }
-  return ranges;
-};
+import { generatePastWeekDateRanges } from "@/lib/dateRange";
 
 const pastWeekDateRanges = generatePastWeekDateRanges();
