+++ booking-admin/src/app/(protected)/sports-and-fitness/sports/page.tsx
@@ -5,25 +5,9 @@ import { OverviewCardsSkeleton } from "../_components/overview-cards/skeleton";
 import { Suspense, useEffect, useState } from "react";
 import { OverviewCardsGroup } from "../_components/overview-cards";
 import { handleMainRecord } from "@/utils/ReportHelpers/mainHandle";
+import { generatePastWeekDateRanges } from "@/lib/dateRange";
 
 const GymMembershipTable = dynamic(() => import("../_components/gym-membership-table"), { ssr: false });
-const generatePastWeekDateRanges = () => {
-  const now = new Date();
-  const ranges = [];
-  for (let i = 0; i <= 7; i++) {
-    const pastDate = new Date(now.getTime() - i * 24 * 60 * 60 * 1000);
-    ranges.push({
-      label:
-        i === 0
-          ? `Today (${pastDate.toLocaleDateString()})`
-          : i === 1
-          ? `Yesterday (${pastDate.toLocaleDateString()})`
-          : `${i} days ago (${pastDate.toLocaleDateString()})`,
-      value: pastDate.toISOString().split("T")[0],
-    });
-  }
-  return ranges;
-};
 
 const pastWeekDateRanges = generatePastWeekDateRanges();
