+++ booking-admin/src/hooks/useOrganisationInfo.ts
@@ -1,8 +1,7 @@
 'use client'
-/* eslint-disable */
 import { useEffect, useState } from "react";
+import { getOrganisationInfo } from "@/lib/getOrganisationInfo";
 import { defaultOrganisationInfo } from "@/config/settings";
-import { organisationInfoEndpoints } from "@/utils/dataEndpoint/organisation-info";
 
 export function useOrganisationInfo() {
   const [organisation, setOrganisation] = useState(defaultOrganisationInfo);
@@ -10,30 +9,24 @@ export function useOrganisationInfo() {
   const [error, setError] = useState<string | null>(null);
 
   useEffect(() => {
+    let isMounted = true;
     async function fetchOrganisation() {
       setLoading(true);
       try {
-        const apiData = await organisationInfoEndpoints.getOrganisationInfos();
-        let merged = { ...defaultOrganisationInfo };
-        if (Array.isArray(apiData) && apiData.length > 0) {
-          // If API returns logo as string, convert to object for compatibility
-          let apiLogo = apiData[0].logo;
-          if (typeof apiLogo === "string") {
-            apiLogo = { light: apiLogo, dark: apiLogo };
-          } else if (!apiLogo) {
-            apiLogo = { light: merged.logo.light, dark: merged.logo.dark };
-          }
-          merged = { ...merged, ...apiData[0], logo: { ...merged.logo, ...apiLogo } };
-        }
-        setOrganisation(merged);
+        // Always use getOrganisationInfo to fetch and merge API data with defaults
+        const { organisation: apiOrganisation } = await getOrganisationInfo();
+        if (isMounted) setOrganisation(apiOrganisation);
       } catch (err: any) {
-        setError(err.message || "Failed to fetch organisation info");
-        setOrganisation(defaultOrganisationInfo);
+        if (isMounted) {
+          setError(err.message || "Failed to fetch organisation info");
+          setOrganisation(defaultOrganisationInfo);
+        }
       } finally {
-        setLoading(false);
+        if (isMounted) setLoading(false);
       }
     }
     fetchOrganisation();
+    return () => { isMounted = false; };
   }, []);
 
   return { organisation, loading, error };
